--- controllers/provisioners/eks/types.go
@@ -1,68 +1,0 @@
-/*
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package eks
-
-import (
-	"github.com/keikoproj/instance-manager/api/v1alpha1"
-	"github.com/keikoproj/instance-manager/controllers/common"
-	"github.com/keikoproj/instance-manager/controllers/providers/aws"
-)
-
-type EksDefaultConfiguration struct {
-	DefaultSubnets []string `yaml:"defaultSubnets,omitempty"`
-	EksClusterName string   `yaml:"defaultClusterName,omitempty"`
-}
-
-type EksInstanceGroupContext struct {
-	InstanceGroup    *v1alpha1.InstanceGroup
-	KubernetesClient common.KubernetesClientSet
-	AwsWorker        aws.AwsWorker
-	DiscoveredState  *DiscoveredState
-	ControllerRegion string
-}
-
-func (ctx *EksInstanceGroupContext) GetInstanceGroup() *v1alpha1.InstanceGroup {
-	if ctx != nil {
-		return ctx.InstanceGroup
-	}
-	return &v1alpha1.InstanceGroup{}
-}
-
-func (ctx *EksInstanceGroupContext) GetUpgradeStrategy() *v1alpha1.AwsUpgradeStrategy {
-	if &ctx.InstanceGroup.Spec.AwsUpgradeStrategy != nil {
-		return &ctx.InstanceGroup.Spec.AwsUpgradeStrategy
-	}
-	return &v1alpha1.AwsUpgradeStrategy{}
-}
-
-func (ctx *EksInstanceGroupContext) GetState() v1alpha1.ReconcileState {
-	return ctx.InstanceGroup.GetState()
-}
-
-func (ctx *EksInstanceGroupContext) SetState(state v1alpha1.ReconcileState) {
-	ctx.InstanceGroup.SetState(state)
-}
-
-func (ctx *EksInstanceGroupContext) GetDiscoveredState() *DiscoveredState {
-	if ctx.DiscoveredState == nil {
-		ctx.DiscoveredState = &DiscoveredState{}
-	}
-	return ctx.DiscoveredState
-}
-
-func (ctx *EksInstanceGroupContext) SetDiscoveredState(state *DiscoveredState) {
-	ctx.DiscoveredState = state
-}
